requirements.txt
# SENKRON Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Astronomical calculations
pyephem>=4.1.5

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # FastAPI test client için

# Code Quality
ruff>=0.1.0

# Optional: Production extras
gunicorn>=21.2.0
